Call init_chem when config_init_test_species is true
The init_chem routine was previously called unconditionally, but that call was
within a preprocessing directive that used the CHEMISTRY macro, which is never
defined. Consequently, the init_chem routine never actually got called.

Now, the call to the init_chem routine depends on the config_init_test_species
namelist option, and no preprocessing directives are used to control the call to
this routine.